third_party/blink/web_tests/platform/generic/fast/events/constructors/speech-recognition-error-constructor-expected.txt

This tests the constructor for the SpeechRecognitionError DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new webkitSpeechRecognitionError('eventType').bubbles is false
PASS new webkitSpeechRecognitionError('eventType').cancelable is false
PASS new webkitSpeechRecognitionError('eventType').error is ""
PASS new webkitSpeechRecognitionError('eventType').message is ""
PASS new webkitSpeechRecognitionError('eventType', { bubbles: false }).bubbles is false
PASS new webkitSpeechRecognitionError('eventType', { bubbles: true }).bubbles is true
PASS new webkitSpeechRecognitionError('eventType', { cancelable: false }).cancelable is false
PASS new webkitSpeechRecognitionError('eventType', { cancelable: true }).cancelable is true
PASS new webkitSpeechRecognitionError('eventType', { error: 'bar' }).error is "bar"
PASS new webkitSpeechRecognitionError('eventType', { message: 'foo' }).message is "foo"
PASS successfullyParsed is true

TEST COMPLETE
